Remove all usage of a canvas element
This (major) change is an effort to make the library as flexible and future-proof as possible. Instead of needing a canvas to draw to, and having to handle the smoothing, shadow*, etc. options, it now simply returns each rendered frame as an `ImageBitmap` (via a promise) that the user can then paint to their canvas via `drawImage()`. The user now has full control over their canvas and its options (or doesn't have to use a canvas at all!)